support for mbed log v2 and refactors related to it
