+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.19.0]
+
+### Added
+
+- Support for Mbed Log version 2
+
 ## [0.18.4]
 
 ### Changed
