+++ Cargo.toml
@@ -2,7 +2,7 @@
 name = "logviewer-rs"
 description = "Log viewer app for viewing plots of data from projects such as motor and generator control"
 authors = ["SkyTEM Surveys", "Marc Beck König"]
-version = "0.18.4"
+version = "0.19.0"
 edition = "2021"
 repository = "https://github.com/luftkode/logviewer-rs"
 homepage = "https://github.com/luftkode/logviewer-rs"
@@ -28,6 +28,8 @@ strum = "0.26.3"
 strum_macros = "0.26.3"
 byteorder = "1.5.0"
 chrono = { version = "0.4.38", features = ["serde"] }
+getset = "0.1.3"
+derive_more = { version = "1", features = ["full"] }
 
 # Dev dependencies
 testresult = "0.4.1"
