+++ crates/skytem_logs/src/mbed_motor_control.rs
@@ -1,5 +1,4 @@
 pub mod mbed_config;
-pub mod mbed_header_v1;
-pub mod mbed_header_v2;
+pub mod mbed_header;
 pub mod pid;
 pub mod status;
