+++ src/app/supported_logs.rs
@@ -4,9 +4,15 @@ use serde::{Deserialize, Serialize};
 use skytem_logs::{
     generator::{GeneratorLog, GeneratorLogEntry},
     mbed_motor_control::{
-        pid::{header_v1::PidLogHeaderV1, PidLog},
-        status::{header_v1::StatusLogHeaderV1, StatusLog},
-        MbedMotorControlLogHeader,
+        mbed_header::MbedMotorControlLogHeader,
+        pid::{
+            header_v1::PidLogHeaderV1, header_v2::PidLogHeaderV2, pidlog_v1::PidLogV1,
+            pidlog_v2::PidLogV2,
+        },
+        status::{
+            header_v1::StatusLogHeaderV1, header_v2::StatusLogHeaderV2, statuslog_v1::StatusLogV1,
+            statuslog_v2::StatusLogV2,
+        },
     },
 };
 use std::{
@@ -19,8 +25,10 @@ use std::{
 /// that would require the log interface to also support a common way for plotting logs
 #[derive(Default, Deserialize, Serialize)]
 pub struct SupportedLogs {
-    pid_log: Vec<PidLog>,
-    status_log: Vec<StatusLog>,
+    pid_log_v1: Vec<PidLogV1>,
+    pid_log_v2: Vec<PidLogV2>,
+    status_log_v1: Vec<StatusLogV1>,
+    status_log_v2: Vec<StatusLogV2>,
     generator_log: Vec<GeneratorLog>,
 }
 
@@ -28,10 +36,16 @@ impl SupportedLogs {
     /// Return a vector of immutable references to all logs
     pub fn logs(&self) -> Vec<&dyn Plotable> {
         let mut all_logs: Vec<&dyn Plotable> = Vec::new();
-        for pl in &self.pid_log {
+        for pl in &self.pid_log_v1 {
             all_logs.push(pl);
         }
-        for sl in &self.status_log {
+        for pl in &self.pid_log_v2 {
+            all_logs.push(pl);
+        }
+        for sl in &self.status_log_v1 {
+            all_logs.push(sl);
+        }
+        for sl in &self.status_log_v2 {
             all_logs.push(sl);
         }
         for gl in &self.generator_log {
@@ -43,8 +57,10 @@ impl SupportedLogs {
     /// Take all the logs currently store in [`SupportedLogs`] and return them as a list
     pub fn take_logs(&mut self) -> Vec<Box<dyn Plotable>> {
         let mut all_logs: Vec<Box<dyn Plotable>> = Vec::new();
-        all_logs.extend(self.pid_log.drain(..).map(|log| log.into()));
-        all_logs.extend(self.status_log.drain(..).map(|log| log.into()));
+        all_logs.extend(self.pid_log_v1.drain(..).map(|log| log.into()));
+        all_logs.extend(self.pid_log_v2.drain(..).map(|log| log.into()));
+        all_logs.extend(self.status_log_v1.drain(..).map(|log| log.into()));
+        all_logs.extend(self.status_log_v2.drain(..).map(|log| log.into()));
         all_logs.extend(self.generator_log.drain(..).map(|log| log.into()));
 
         all_logs
@@ -106,12 +122,25 @@ impl SupportedLogs {
     // Parsing dropped content on web
     fn parse_content(&mut self, mut content: &[u8]) -> io::Result<()> {
         if PidLogHeaderV1::is_buf_header(content).unwrap_or(false) {
-            self.pid_log.push(PidLog::from_reader(&mut content)?);
+            let log = PidLogV1::from_reader(&mut content)?;
+            log::debug!("Got: {}", log.descriptive_name());
+            self.pid_log_v1.push(log);
         } else if StatusLogHeaderV1::is_buf_header(content).unwrap_or(false) {
-            self.status_log.push(StatusLog::from_reader(&mut content)?);
+            let log = StatusLogV1::from_reader(&mut content)?;
+            log::debug!("Got: {}", log.descriptive_name());
+            self.status_log_v1.push(log);
+        } else if PidLogHeaderV2::is_buf_header(content).unwrap_or(false) {
+            let log = PidLogV2::from_reader(&mut content)?;
+            log::debug!("Got: {}", log.descriptive_name());
+            self.pid_log_v2.push(log);
+        } else if StatusLogHeaderV2::is_buf_header(content).unwrap_or(false) {
+            let log = StatusLogV2::from_reader(&mut content)?;
+            log::debug!("Got: {}", log.descriptive_name());
+            self.status_log_v2.push(log);
         } else if GeneratorLogEntry::is_bytes_valid_generator_log_entry(content) {
-            self.generator_log
-                .push(GeneratorLog::from_reader(&mut content)?);
+            let log = GeneratorLog::from_reader(&mut content)?;
+            log::debug!("Got: {}", log.descriptive_name());
+            self.generator_log.push(log);
         } else {
             return Err(io::Error::new(
                 io::ErrorKind::InvalidData,
@@ -125,16 +154,29 @@ impl SupportedLogs {
     fn parse_path(&mut self, path: &path::Path) -> io::Result<()> {
         if PidLogHeaderV1::file_starts_with_header(path).unwrap_or(false) {
             let f = fs::File::open(path)?;
-            self.pid_log
-                .push(PidLog::from_reader(&mut BufReader::new(f))?);
+            let log = PidLogV1::from_reader(&mut BufReader::new(f))?;
+            log::debug!("Got: {}", log.descriptive_name());
+            self.pid_log_v1.push(log);
         } else if StatusLogHeaderV1::file_starts_with_header(path).unwrap_or(false) {
             let f = fs::File::open(path)?;
-            self.status_log
-                .push(StatusLog::from_reader(&mut BufReader::new(f))?);
+            let log = StatusLogV1::from_reader(&mut BufReader::new(f))?;
+            log::debug!("Got: {}", log.descriptive_name());
+            self.status_log_v1.push(log);
+        } else if PidLogHeaderV2::file_starts_with_header(path).unwrap_or(false) {
+            let f = fs::File::open(path)?;
+            let log = PidLogV2::from_reader(&mut BufReader::new(f))?;
+            log::debug!("Got: {}", log.descriptive_name());
+            self.pid_log_v2.push(log);
+        } else if StatusLogHeaderV2::file_starts_with_header(path).unwrap_or(false) {
+            let f = fs::File::open(path)?;
+            let log = StatusLogV1::from_reader(&mut BufReader::new(f))?;
+            log::debug!("Got: {}", log.descriptive_name());
+            self.status_log_v1.push(log);
         } else if GeneratorLog::file_is_generator_log(path).unwrap_or(false) {
             let f = fs::File::open(path)?;
-            self.generator_log
-                .push(GeneratorLog::from_reader(&mut BufReader::new(f))?);
+            let log = GeneratorLog::from_reader(&mut BufReader::new(f))?;
+            log::debug!("Got: {}", log.descriptive_name());
+            self.generator_log.push(log);
         } else {
             return Err(io::Error::new(
                 io::ErrorKind::InvalidData,
@@ -177,18 +219,18 @@ fn is_zip_file(path: &Path) -> bool {
 mod tests {
     use super::*;
     const TEST_DATA_STATUS: &str =
-        "test_data/mbed_motor_control/20240926_121708/status_20240926_121708_00.bin";
+        "test_data/mbed_motor_control/v1/20240926_121708/status_20240926_121708_00.bin";
 
     const TEST_DATA_PID: &str =
-        "test_data/mbed_motor_control/20240926_121708/pid_20240926_121708_00.bin";
+        "test_data/mbed_motor_control/v1/20240926_121708/pid_20240926_121708_00.bin";
 
     #[test]
     fn test_supported_logs_dyn_vec() {
         let data = fs::read(TEST_DATA_STATUS).unwrap();
-        let status_log = StatusLog::from_reader(&mut data.as_slice()).unwrap();
+        let status_log = StatusLogV1::from_reader(&mut data.as_slice()).unwrap();
 
         let data = fs::read(TEST_DATA_PID).unwrap();
-        let pidlog = PidLog::from_reader(&mut data.as_slice()).unwrap();
+        let pidlog = PidLogV1::from_reader(&mut data.as_slice()).unwrap();
 
         let v: Vec<Box<dyn Plotable>> = vec![Box::new(status_log), Box::new(pidlog)];
         assert_eq!(v.len(), 2);
