+++ test_data/mbed_motor_control/v1/20240926_121708/status_20240926_121708_00.bin
